+++ Assets/Scripts/Tools/HealthBarManager.cs
@@ -11,7 +11,7 @@ public class HealthBarManager : MonoBehaviour
     {
         if(linkedEntity.maxHp != 0)
         {
-            float size = Mathf.Clamp01(linkedEntity.currentHp / linkedEntity.maxHp);
+            float size = Mathf.Clamp01((float)linkedEntity.currentHp / linkedEntity.maxHp);
             bar.transform.localScale = new Vector3(size, 1, 1);
         }
